MBS-11862: Do not show deprecated relationship types with 0 uses
We've historically just removed link types once we no longer wanted
them used and they had 0 uses, but that is a bad idea. It means
the historical edits automatically show less info, since there's no type
to load anymore, and it's just an unwanted loss of historical data.
As such, this allows us to keep them in the database, but not show them
except when there's something to gain from doing so:
for type editing / documenting, and for the edit search.