Fix deprecated link type code + tests
This test broke with ec722a87b939d4d113e8b44ee1c5bb17c981f39c
because MusicMoz, having zero uses, is now not shown for selection.
This adds one use for that to make sure it can still be found,
and a description so that we don't need to artificially enable
the selector like we were doing.
It also adds one new link type which is deprecated and has zero uses,
to test it doesn't show up at all.

Once the test was fixed, it showed that there *was* also a bug with
the implementation - it was seemingly comparing against the wrong
id in $extra_condition. That should work now.