MBS-11693: Give useful message when rejecting Musixmatch /album links (#2126)
When we decided to block these we didn't yet have a way
to indicate why we were blocking it in the error message,
but we specifically suggested in the forums we should add that info
once it was possible. Then... we forgot about it. This adds that info.
I checked in case the situation got better in the meantime
and we should allow the links instead, but it doesn't seem
to be the case.