MBS-11823: Don't break tags and genres mid-word (#2214)
This was set to word-break: break-all to avoid long one-word tags
such as weird URLs. That actually means all tags with normal words
break weirdly. mwiencek proposed overflow-wrap: anywhere to cater
to both of these issues, and that seems to work - the only worry
is that this is not supported in older browsers.