MBS-9902(I): Restrict link type options
Support multiple roles of a URL.

Pass typeOptions as an array of objects instead of HTML elements
so that it can be filtered programmatically,
type definition by @reosarevok.

Add a new 'getPossibleTypes' function in URLCleanup
and a 'filterTypeOptions' function in externalLinks
for restricting type options to achieve better UX.

Remove deprecated link types detection tests
since those types are filtered after restriction.

Update seeding tests because ASIN is not available for a random link.