MBS-11866: Interface for mass deletion of accounts
Account admins could sometimes use a way to remove a lot of accounts
doing the same thing (for example, spammily rating or tagging
an entity). This allows them to do that, by giving a bunch of editor
names. This gets turned into a list of editors with more data,
in order to check that the editors should in fact get deleted,
after which the account admin can proceed to the deletion.

This does not allow account name reuse because from talking to
Freso, it's almost never desired for a case like this, and if it
ever is then the usernames can be unlocked elsewhere in the admin tools.