Allow disabling the template renderer
If `RENDERER_SOCKET` is empty, MBS will now avoid communicating with the
template renderer at all. This will prevent socket timeouts ("Couldn't
connect to the renderer") from occurring if certain non-WS requests leak
into our WS containers in production (as these containers don't have any
renderer process).

A blank response is returned if this happens, because it's not an error
that's even intended to be /allowed/ to be triggered normally (and I
don't suspect anyone will use this flag outside of us), but I improved
the logging to include the URL that triggered it. (You can already find
which URLs are causing this in Sentry.)