MBS-10902: Stop changing zxx language name at the Perl level
This was causing issues when both work and release edits
were being loaded, in that releases would appear with the language
[No lyrics]. Now that all relevant edits are React anyway,
there's no need to massage the work on the Perl level at all.
Instead, this just passes the languages (rather than their names)
to JS and lets localizeLanguageName change the displayed name only.