MBS-11948: More strict favicon class assignment
We have had several issues in the past of favicons being assigned
incorrectly. The one triggering this issue was kenjimiura.com
being matched with the residentadvisor "ra.co" string.
There's no reason why we should stick to this naive approach to matching
when we have regular expressions and URL objects, so this
makes use of both to actually validate the full hostnames of URLs
and, when needed, the start of their paths.
The regexes are mostly based on the existing matchup/cleanup
we do. The few we don't validate/clean up at all I based
on what the structure seems to be via a quick database check,
although admittedly I might have missed some edge cases.

As an additional bonus, I added testing for the favicon classes,
to make sure that at least all the URLs we were already testing
do match the expected favicon class (that should also serve
as a reminder to change the regexes there if we change how a site
is cleaned up).