MBS-11799: Display "credited as" field in sidebar external links (#2180)
It makes sense to specify the credits for URL rels on the sidebar.
On one hand, it helps find them to fix them when they are
unwanted results from a merge. On the other hand, it shows
the differences between, say, several Discogs pages for the same
MB artist under different aliases.

Making the whole thing one exp.l because having {url} as {credit}
seems a lot easier to translate.