MBS-6140: Allow (recording|work)-level-rels for release browse (#2173)
This was apparently not implemented because of fears of too big
release lists causing load issues, but since we have a max_recording_relationships
limit now, that shouldn't be a worry anymore. As such, this saves users
one call per release on most cases (except if they'd request over 500 recordings).