Keep filter when navigating with ArtistIndex footer
This makes it so that when we move from one ArtistIndex section
to another (with the FooterSwitch) with a filter on, most of the
filter settings are kept, since they apply to all pages
(there's no reason to drop the name or type search just because
the user moved from this artist to VA).

Artist credit filters are kept, but only work if the destination
filter also has the same artist credit (so, they always work
when moving from official to all, sometimes when moving from
all to official, and never when moving from single to va or back).
The filter stays in the URL when moving but it's just ignored
if it doesn't match an available artist credit.