Allow limiting compression threads from config
To compress JSON dumps (using xz) and search index dumps (using zstd),
the number of threads is currently not limited so dumps can be made
available faster. However, it now has detrimental impacts to the
server’s availability when dumps are built.

This patch adds the key DUMP_COMPRESSION_THREADS to the config so it can
be changed at deployment time. It doesn’t change the default though.