Add seeded links after any existing ones
We were first loading the seeded links, then sorting
all the initial links. But that puts the newly added links
in between any existing ones, which is confusing
(as mentioned in MBS-11960).
This changes it so that first we sort the existing links,
and then we add any links being seeded, so that
they appear at the bottom, where new links usually are.
That has the extra benefit of making it so that the seeded links
are shown in the order they were seeded in, as far as I can tell.