MBS-10056: Don't add redundant credit to relationship on entity merge
When two entities that can have relationship credits
are merged and relationship credits haven't been
chosen for renaming, the name of the old entity is added
as a credit for any relationships it had that had
no credit already, in order to save the name.

This makes sense in most cases, but it does not make sense
for the merge of two entities of the same name
(but with different disambiguation or whatever).
In that case, we end up with a relationship with a credit
that already matches the merged entity name.
As such, it just looks a bit dumb and is not required at all
to save the name, which is the only reason we're doing this
to begin with. This makes sure the name is not the same
before saving it as a credit.