Timestamp every line of log from cron jobs
Scripts running as cron jobs do timestamp some messages but not all of
these, and - more problematic - they do not timestamp error messages.

Cron command was already redirecting to the log file. It now timestamps
every line of log using the command `ts` from the package `moreutils`.

The added wrapper script `run` is meant to be run in production
containers such as json-dump, production-cron, search-indexes-dump,
and sitemaps. It can be used for both cron tasks and occasional runs.
It makes sure to run using the user `musicbrainz` as required on the
primary MusicBrainz server; That simplifies invoking scripts in
production. Otherwise it runs using the current user to stay compatible
with mirrors and development servers that may not use the user
`musicbrainz` for now.

Given a command to be run, it:
1. Sets the environment variables PATH and PERL_CARTON_PATH;
2. Sets the current working directory to the MusicBrainz Server folder;
3. Checks the log subdirectory corresponding to the command;
4. Sudo to the user `musicbrainz` if available and not already current;
5. Executes the command using Perl `carton`;
6. Timestamp and log the output to a log file.