MBS-11977 Fix targeting ENTITY for URL rel error
When an external website can match only one auto-selected relationship
type name (for a given entity type), if a URL matches the overall
pattern for such external website but doesn’t follow the specific
pattern for entity type-specific relationship type, that means it should
rather be linked to an entity of another type. `ERROR_TARGETS.ENTITY` is
the appropriate target in this case.

This patch fixes most `validate` functions but the ones that would
require a larger refactoring to determine the exact target.