Change EOSQL to SQL for TextMate/VSCode
TextMate (so, VSCode, possibly other IDEs too?) recognize
Perl heredocs with "SQL" as a SQL block, allowing them to be
highlighted and parsed as such. That makes finding issues
a lot easier, so I'm changing our uses of EOSQL to SQL with a quick
find and replace.
This should be easily customizable in vi and whatnot
so it shouldn't cause serious issues to anyone else, hopefully.