MBS-11979: Change example URL to example.com
The previous example `google.com` was a valid URL but search pages are
not actually wanted to link to any MB entity.

This patch replaces it with `example.com` to get rid of this ambiguity.
It also rephrases to skip repeating “example”, to be kind, and to format
the URL the same way it is formatted in `URLCleanup.js` already.