MBS-11852: Only call TO_JSON on editor if it exists (#2246)
When trying to approve an edit with a No vote + note, we load edit notes
for the edit (to see if one of the editor_ids matches the approver's)
but we do not load the actual editors themselves.
On doing TO_JSON for edit when passing it to NoteIsRequired,
we were (correctly) running to_json_array on the edit notes, and this
was running TO_JSON directly on the note's editor, which, of course,
had not been loaded. This changes it to run to_json_object,
which is smarter and actually ensures TO_JSON can be called
before calling it.