MBS-11909: Show user's own collections in collaborator's page
If a user owns a private collection, they should be able
to see it listed in the collections page of any of the collaborators
in that collection - it doesn't make sense to keep it private
for the collection owner!

Collection->find_by is still a mess of spaghetti
but in this case it seems to be more than good enough
to just have one condition
for collaborations here.