MBS-11909: List viewable private collections under subscriptions
If a user is subscribed to a private collection, be it
their own or one they collaborate on, they should be able
to see it listed under their collection subscriptions.
Also, the owner and other collaborators of a collaborative
collection should be able to see the subscription here,
since the subscription itself is not private - the collection is,
and they have rights to see it.