MBS-11975: Clean up CDJapan detailview URLs
CDJapan has detailview.html URLs that seem to just redirect
to the product ones, but have been added a few times
and are not currently detected at all by the code.
This changes the CDJapan autoselect to detect them
and clean them to their final product page form.

Since this takes care of the only kind of http URLs
we still were getting for CDJapan, it should be safe
to remove the http override too (existing ones are being fixed by hand).