MBS-11989: Also correct rg to release_group for delete_alias
We already do this for add_alias and edit_alias,
but I guess we forgot to do it for delete_alias,
and that is causing an error when loading the layout on React.
Why we use "rg" I don't know, but in any case it seems
sensible to have all of these work until we maybe decide to change it.