MBS-11995: Test error messages in external links editor
We should ensure we are returning the expected errors
for tests that we expect to fail, including both specific message
(when given) and target.
This just uses a substring, rather than the whole error message,
because it seems like overkill otherwise and it's easier for ones
we expand with arguments.

Adds a new optional test property 'expected_error'.
This requires one of 'input_relationship_type' or
'expected_relationship_type' (since the error might vary
depending on the relationship selected) and 'input_entity_type'
(since the same URL might be allowed for one entity type but not
another).

This adds two examples:

One for our block of Wikipedia pages as discography entries for
releases, testing a specific error message and
'input_relationship_type'.

One for blocking non-specific 45cat pages, testing a generic
error message (undefined error) and 'expected_relationship_type'.

Further tests to cover other existing blocks and errors
will be added in subsequent commits.