MBS-11952: Autoselect Vimeo On Demand URLs correctly (#2262)
Vimeo has a Vimeo On Demand service which allows video makers
to rent out and sell their videos. These give a smaller rental price
for 24 hours, allowing you to stream from the website,
and a larger price to purchase outright, allowing streaming
and downloads.

It seems in a few cases only one or the other are available,
so rather than autoselecting, this just restricts the options
to one (or both) of those, for release or recording only.

This blocks /store/ondemand links (the ones you get when trying
to actually rent or buy the video) since there's no automatic
way to turn them back into the main video URL that I can see.

To make our naive favicon implementation use the ondemand
favicon rather than the main vimeo one, I set the lines
in the wrong sort order. I'm hoping to refactor this soon
and get rid of this constant completely, so it shouldn't be too bad.