Detect whether to reset the Selenium env
reset_selenium_env.sh typically takes anywhere from 10-15 seconds to
run. That's a lot of time when multiplied by the number of tests.

A good number of tests don't actually make any changes to the DB,
though, so we can skip this step in that case. We detect whether any
changes were made via a new script, get_selenium_tup_stats.sh.

One issue that had to be resolved was that logging in updates the
last_login_date in the database. I've added a new DBDefs flag to avoid
this (which is also useful when connecting to a standby).