Consolidate List::UtilsBy uses into List::AllUtils
The whole point of List::AllUtils is that it combines
List::UtilsSome, List::UtilsBy and List::Util. As such,
this uses List::AllUtils instead of List::UtilsBy.
That actually means it will use the sort_by and nsort_by
functions of SomeUtils instead, but that is equivalent
except when called in scalar context. We only had one
issue of that, and it seemed to be an unintentional bug to begin with
(since calling UtilsBy nsort_by in scalar context returns the last
element of the list, but the code seemed to want the first).