Use List::AllUtils instead of List::MoreUtils
I checked the source for both, and all of List::MoreUtils methods here
are also on List::SomeUtils with exactly the same code
(except some style changes with spacing and whatnot),
and List::SomeUtils is part of List::AllUtils.
There is one exception, pairwise, which works in a slightly
different way which is slightly problematic in PP but fine in XS.
As far as mwiencek and I could tell, it should be safe to use for us.