Replace ProhibitUnusedVariables with ProhibitUnusedVarsStricter
https://metacpan.org/pod/Perl::Critic::Policy::Variables::ProhibitUnusedVarsStricter
Variables::ProhibitUnusedVarsStricter is a more strict (surprise)
version of ProhibitUnusedVariables.

Turns out we have a ton of unused variables.

On ($var, $unused_var) = @x assignations, I'm leaving undef instead
of the unused variable when that seems clearer, removing it if not.

For lexically scoped wrappers, I added some documentation to make it
more clear, and I explicitly destroy them both for clarity
and to make them not be detected as unused to begin with.