MBS-11985: Process seeded relationships on the server (#2284)
Doing this on the client meant any seeded target entities had to be
requested asynchronously before we could display the relationships. On
the server, we can just load the target entities from the database.

Adds a selenium test. (There was already a small one in
t/selenium/MBS-11730.json5, but it only tested one seed parameter and
didn't test submitting the relationships.)