MBS-12018: Use /g for post process words that might repeat
There's no reason why we can't have "My Favourite DJs Are All DJs"
or "Call Me on My iPhone (That's What an iPhone Is For)",
so we should process *all* matches of the word in a string
when post-processing.
I left /g off for things that would be expected to appear
only once in brackets, such as 're-edit', or at all, such as 'f/'.