MBS-11974: Fix track relationship grouping for attributes/dates (#2281)
This is a follow-up to MBS-11806.

Relationships with different attributes or dates should not be grouped
into a single track range.

For example, if we have the relationships:

  phonographic copyright by: Aniplex Inc. (in 2007) (track 1)
  phonographic copyright by: Aniplex Inc. (in 2008) (track 2)

we are currently grouping these as:

  phonographic copyright by: Aniplex Inc. (in 2007, in 2008) (tracks 1–2)

which is not ideal, because it's ambiguous which tracks' copyrights are
in 2007 and which are in 2008. (But the bug may apply to any type of
relationship.)

Instead of unconditionally appending tracks to every similar target
group, we first ensure that target groups are not merged for different
tracks. Next, once we've run `mergeDatedExtraAttributes` on each target
group, we merge ones with equal date/attribute lists, combining their
track sets. This is done in `mergeTargetGroupsByTracks`.

The way we generate a key for each `RelationshipTargetGroupT` (which was
intended as a fast way to perform a comparison, and to be used for the
React key prop) became incomplete, because it didn't include the list of
tracks. I think generating a fully-representative key with tracks would
be slower and more complex than just giving it a unique ID and running
the comparator function again, so I've done that instead.