Hydrate the TaggerIcon with React
As opposed to the current method of running a snippet of jQuery on every
single page. We shouldn't be stuffing random things into common.js: it
increases the amount of code that needs to be downloaded and executed on
all pages.

I've changed the props for TaggerIcon from the recording/release object
to just the entityType and gid. This limits the amount of JSON needed to
embed for hydration.

The commit message from e1ee68bea86 has been copied into the file mostly
as-is for clarity.

I removed the `&t=$timestamp` part of the tagger link query string
because it's not able to be made identical for hydration on the client
and server. While testing against Picard, it didn't appear to affect
cache busting at all, perhaps because Picard returns `Cache-Control:
max-age=0` on the response already.