Move $entity.js scripts to $entity/edit.js
The path "scripts/$entity.js" doesn't indicate which page the script is
for. In every case I can see, it's for the edit page. I'd rather they be
moved to "scripts/$entity/edit.js", then, because as we add scripts for
other entity sub-pages, it'd be odd for the edit page script to be
top-level above other ones.