Improve AcoustIdCell implementation
The previous implementation meant that (due to how hydration works) any
page using the RecordingList component had to embed the entirety of its
props (including all of the recordings being listed) as JSON in the
HTML. This was especially not ideal because the only bits of
RecordingList needing hydration, the AcoustIdCells, aren't displayed
anywhere outside of the recording merge page.

This moves hydration (and thus data fetching) into the AcoustIdCell
component, with some changes to continue to allow requests to the
AcoustID API to batch together.

Another minor issue I've fixed is the lack of translation of the
'AcoustIDs' header on the client. This is done by simply moving the
column definition to be under root/static/scripts/.