MBS-8570: Check for duplicates based only on current links
We shouldn't check for duplicates based on both kept and deleted links,
since that means the duplicate is still detected even when deleting
an already existing dupe.

The reason we did this was to avoid an editor deleting a relationship,
then adding the exact same relationship. We can still block that
by checking it separately only if the link is new or changed,
and actually giving a more useful error message than "already exists"
that explicitly tells the user to prefer the existing relationship
they are deleting.