MBS-12001: Use allowNew for recording links when adding a new track
For some reason, we did this check for track changes
(so, when a new recording is assigned to an existing track)
but not for new track adds that create a new recording.

This is a visual change only - the edits do the right thing either
way, but this will avoid the recording showing as deleted until
it gets created.