Fix last react-hooks/exhaustive-deps issues
`useCallback` isn't needed for callbacks passed to HTML elements like
`<a>`. I removed it from those places.

The issue in AcousticBrainz.js was on `useEffect`, not `useCallback`,
though I had to add `useCallback` to one callback in order to make it
stable when passed to `useEffect`.

I changed this rule from warn to error, because sometimes it does
indicate a real issue with the code.