Store all edit type ids as both constants and types
We'll need all of these for edit details typing later.

Flow is annoying, so we can't import the constants to use the typing.
As such, I'm just defining the types and the constants separately,
which is even more duplication, but flow should ensure we keep the
two in sync anyway.