MBS-12059: Autoselect YesAsia as "purchase for mail-order" + sidebar
The URLs do seem to have somewhat different patterns for artist
(possibly also label?) vs product, but I didn't find clear documentation
and they're not obvious enough that I'd dare to guess. As such, I'm just
checking whether there's info (for releases) or list (for artists)
on the (info|list).html section at the end of the URL for now, since that
does seem to be consistent enough.