MBS-7727: Restrict Soundcloud to only "get the music" rels
Soundcloud shouldn't be used for license nor crowdfunding
relationships (nor, probably, any other that gets added
further down the line). This restricts it to only allow
"get the music" relationships (all of them, because apparently
some stuff on SoundCloud requires a SoundCloud Go+ subscription,
whatever that involves  - not available in my country).

For now, this allows only the rels as singletons or as free+free
or paid+paid combinations. If other valid options are found,
we can add them too.