MBS-12047: Support browsing artist RGs in /ws/2 as in website index
This is useful to be able to easily get through the API
the same list of release groups we show by default
on the MusicBrainz site itself.

I chose to use a new 'release-group-status' parameter
rather than the existing 'status' parameter
already in use for releases because 'status'
can be used to filter the releases via inc=.
I'm using "website-default" for the option matching the website
since while we call it "official" on the site, is a wider concept than
what we call official for releases (including also for example RGs with
only no-status-set releases) and we might want to allow querying only
RGs containing official releases in the future.