MBS-12055: Allow admins to see private editor-editor subscriptions
Freso asked for an easier way to see if a sockpuppet editor
subscribes to other editors (in order to find more sockpuppets).
They specifically asked for a way to know
the tab contains private data and should not be open when streaming
so I added a class (to be able to hide it) and a warning / highlight.
They also specified they did not want to be able to see
any other subscriptions, so this hides even the summary
from the user subscription page in this case.
Also, since the /artists page is private, we need to specifically
make the tab lead to /editors in this case.