Convert Edit::Show test to use real data
We were doing this with Edit::Cancel already,
so this changes Edit::Show to also use a real edit.
This means we have to drop the test to see if other
closing conditions are shown correctly,
since no current edits use that.
But then, *no current edits use that*.

In exchange, I added a lot more checks here: to see that stuff
that should be hidden / locked when logged out or as a beginnner
is indeed so, to check that some other important parts
of the edit display appear, and to check
that the expected differences when viewing your own edit
indeed happen.