Create generic localizeTypeNameForEntity function
This will be needed for the entity edits headers,
but even as it is, I think it is simpler than having it
on every entity header. For now using it only
where we would otherwise have shown a type,
but not where we have no types to show (url, release, etc).