MBS-11910: Allow editing rel credits in the external links editor
URL relationships can have credits for the non-URL entity, if it
is of a creditable type, but there is no way currently to change
the credit from the entity page (only from the URL /edit page).
This is confusing, and we really should allow editing this data
from the page URL rels are most commonly edited, same as we now
do for relationship dates.
This makes it possible to change the credited entity in the same
popup where we currently change the dates.
Note that this doesn't implement any changes in the release editor
since release-url rels are not currently creditable - if that changes
then make relevant changes to externalLinkRelationship
in root/static/scripts/edit/MB/edit.js