MBS-7912: Add edit search option "edit [not] in subscriptions"
This is one of the last (the last?) options that has a dedicated page, but
cannot be used in edit searches, meaning we can't actually refine the edit
lists on the /subscribed page using edit search.

This search checks (AFAICT) every single condition that /subscribed does:
artist, label and series subscriptions, and all the possible collection
subscriptions. It does *not* check editor subscriptions, which have
a separate page (/subscribed_editors), and can already be searched
within the edit search separately.