Replace && with if/ternary
alastair asked why we use 'conditions && action'
instead of an if statement or ternary in some files,
and both mwiencek and I didn't have any good reasoning
and prefer the alternative option. So this changes
the few cases I could find to use if statements or ternaries.
I feel this makes things a lot more clear.

I also moved some conditions to a constant in index.js
for readability.