MBS-12079: JSON-LD for releases with a lot of [silence] is too big (#2331)
The most-referenced [silence] recording has over 2,000 gid redirects.
These are all outputted in the `sameAs` property in its JSON-LD,
including on release pages that reference it. If the release contains
[silence] multiple times, as
https://musicbrainz.org/release/86427535-9abb-4681-8c88-5c651e55d62a
does for example, then this is multiplied. (The linked release
references [silence] more than 50 times, and contains a 7.75 MB chunk of
JSON-LD.)

It also so happens that Google reports "Structured data with syntax
errors" on this page, with "Incorrect value type" as the reason (but the
part of the source it highlights as having an error makes no sense). So
we're wondering if this is just due to the size of the page hitting some
internal limits that Google has.

This patch limits the number of gid redirects outputted for recordings'
`sameAs` properties on release pages to 100 each. It also skips
outputting them entirely if the recording appears multiple times (only
outputting them for the first occurrence).