MBS-12083: Add caching to get_types_for_timeline
While the absurdly overcomplicated query was the cause of most
of the slowness here, there's still no reason to call PG multiple
times each time someone loads the timeline. This caches the data
for 24 hours - we don't update names often, and if we do, the only
problem we should have is that the name doesn't show in the
timeline for the first 24 hours which doesn't sound
like a big deal.