MBS-9375: Restrict Apple Music to paid streaming and/or purchase
This has been wanted forever, but now the restrict tool
finally allows us to do it.
For now, this just allows paid streaming, paid download or a combination
of both for all entities (artist, label, recording, release).
If we eventually decide to allow for Apple Music relationships
for artists and/or labels, this should be modified accordingly.