Allow specifying the path to temporary directory
Currently, temporary directories are created by scripts under "$TMPDIR"
or '/tmp' using tempdir() of File::Temp in Perl; See references:
* https://metacpan.org/pod/File::Temp#tempdir
* https://metacpan.org/pod/File::Spec#tmpdir

The Perl snippet already shifts the first argument to pass it as
'TMPDIR' - the parent directory of the temporary directory to be
created, but no argument is passed to this snippet.

This patch forwards arguments of the shell function make_temp_dir()
to the Perl snippet so as to allow specifying the parent directory.