MBS-12083: Significantly simplify the get_types_for_timeline query
So, I started working on this using the query to get stats
for each reltype... and it seems I never removed all the useless
(here) joins to link and l_table, which of course
make this a ton slower than it needs to be.
We don't need the count here, all we want is the names,
so this can be a super basic query.