MBS-11312: Add table to store changes to edit notes
While originally I was just planning to make changes
to the edit note table and nothing else, discussion with the team
suggested that it would be better to actually have a table
that stores the changes. This gives us several benefits:
1) we can easily check if an edit note has been deleted or edited
2) we can see who and when made the change
3) we can store the reason the change was made
4) we can see the original value (useful in case a malicious actor
gets access to an admin account somehow, and also useful
to find proof of misconduct if the editor posts an edit note,
then removes it (after it has been emailed to the intended insultee).

This table won't be replicated nor dumped
(other than in private dumps).