MBS-12117: Use CSS specificity to fix error highlighting in RE tabs
This broke when I moved it from release-editor.less
to layout.less because jquery-ui took over.
By specifying the extra classes this takes over
and is again considered more relevant, see
https://developer.mozilla.org/en-US/docs/Web/CSS/Specificity

Also changed the background position for the non-release-editor tabs
because those are smaller and as such the image needs
smaller margins to be centered.