MBS-12056: Convert the edit data page to React (#2316)
This uses formatPluralEntityTypeName as-is rather than
adding a separate function to get lowercase names or
running a lowercasing function on the result of the format function
because I can't see any good reason why the names
need to be lowercase here.
Also, I'm not actively sorting the entity IDs on relatedEntities
(TT used nsort). We could do a number sort here,
but they seem to come in the right order already
and additionally I'm not sure it's even particularly important/useful.

Rather than always passing the raw_data as part of the edit JSON,
this passes it as a separate prop to the component, since we only
use it here. That way we can already pass a pretty-printed JSON
string.