MBS-12124: Add report for license URLs
Most of these are going to be wrong, because people keep
misusing the relationship (even the sample data
already contains a dozen incorrect usages!).
This is sorted with least uses first, rather than last
like other URL reports, since the most used ones are also
the most likely to be correct.