MBS-12126: Add URL search to the site
This is already available via the WS so I see no reason
not to allow users to use it from the site. The results
are kinda mediocre with normal indexed
but still a lot better than not having the option at all,
and return only the correct URL with advanced
and url:'urlstring', so it seems like a no-brainer to me.