MBS-9385: Warn when using non-"None" packaging for digital media
This warns at the medium level mostly because that's what comes last
in the workflow (so it might avoid going back to the first tab
for cases where the error is not the packaging).
I'm not warning on unset packaging because I guess
that's not wrong per se and I'm not sure we want to
pretty much enforce packaging for digital.