MBS-12134: Convert the cancel edit page to React
No significant changes. This reuses EnterEditNote
which means an Edit Note heading is added to that
section - that doesn't seem particularly problematic
so I didn't add a toggle for that.
The TT allowed passing {extra} under the cancel button but
I didn't find any use or reason for it, so this does not allow children.