MBS-11655: Separate tag cloud into genre and non-genre tags
This keeps the existing 200 tag limit for genres, but makes it
the default. Most of the previous cloud was genres,
so it seems fine to keep that as earlier.
It also adds a separate 50 tags cloud for non-genre tags.

Changing the cache key to avoid a clash of the old and new formats
while one is on beta and the other on prod.