Filter empty string IDs in Data::Edit::load_all
We were filtering undef IDs, but not ''.  It appears that some edit
types can return '' in their foreign_keys list.

The foreign_keys are taken from edit data, so some IDs must be stored in
the edit data as ''.  There's not much we can do about that except clean
up the edit data in the future.

Fixes an 'Invalid IDs' warning.