Fix use of uninitialized value in SetCoverArt
The old release ID is undef if the cover art was previously set
automatically. See https://musicbrainz.org/edit/86465803 for example.
Check this before using it as a hash key.

It doesn't appear that the new release ID can ever be undef, but I check
that too for consistency (and it case we ever allow unsetting the
selected artwork).

Fixes the following warning:

"Use of uninitialized value in anonymous hash ({}) at
lib/MusicBrainz/Server/Edit/ReleaseGroup/SetCoverArt.pm line 99."